+++ rust-src/concordium_base/src/pedersen_commitment/randomness.rs
@@ -5,8 +5,6 @@
 
 use crate::{common::*, curve_arithmetic::*};
 
-use ff::Field;
-
 use rand::*;
 use std::ops::Deref;