+++ rust-src/concordium_base/src/sigma_protocols/com_eq_sig.rs
@@ -15,7 +15,6 @@ use crate::{
     ps_sig::{BlindedSignature, BlindingRandomness, PublicKey as PsSigPublicKey},
     random_oracle::RandomOracle,
 };
-use ff::Field;
 use itertools::izip;
 use rand::*;